move code to internal
