+++ internal/agent/app.go
@@ -1,0 +1,12 @@
+package agent
+
+import (
+	"runtime"
+)
+
+func Start() {
+	RuntimeStats = runtime.MemStats{}
+	go collectStats(&RuntimeStats)
+	go reportStats(&RuntimeStats)
+	runtime.Goexit()
+}
