+++ internal/agent/collect.go
@@ -1,4 +1,4 @@
-package main
+package agent
 
 import (
 	"math/rand"
