+++ internal/agent/stats_test.go
@@ -1,4 +1,4 @@
-package main
+package agent
 
 import (
 	"reflect"
