+++ internal/server/storage.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"fmt"
